LC-sqrtx [Runtime: N/A; Memory: N/A]

+++ math/sqrtx.cpp
@@ -4,7 +4,7 @@ public:
         if(x < 1) return 0;
         int left = 1, right = x;
         while(left < right){
-            long long mid = (left + right  + 1) / 2;
+            long long mid = (long long) (left + right  + 1) / 2;
             if(mid * mid > x) right = mid -1 ;
             else left = mid;
         }
